requirements.txt
# Core
Django==2.2
django-configurations==2.1
gunicorn==19.9.0

# For the persistence stores
psycopg2-binary==2.7.7
dj-database-url==0.5.0

# Rest apis
djangorestframework==3.9.1

# Developer Tools
flake8==3.7.5

